Добавлена возможная реализация enum класса для статусов.
